task-4/app/requirements.txt
neo4j==5.14.1
python-dotenv==1.0.0 